+++ pkgs/dart_mcp/lib/src/server/roots_tracking_support.dart
@@ -29,12 +29,6 @@ base mixin RootsTrackingSupport on LoggingSupport {
   /// [_RootsState.pending], otherwise `null`.
   Completer<List<Root>>? _rootsCompleter = Completer();
 
-  /// Whether or not the connected client supports [listRoots].
-  ///
-  /// Only safe to call after calling [initialize] on `super` since this
-  /// is based on the client capabilities.
-  bool get supportsRoots => clientCapabilities.roots != null;
-
   /// Whether or not the connected client supports reporting changes to the
   /// list of roots.
   ///
@@ -44,9 +38,7 @@ base mixin RootsTrackingSupport on LoggingSupport {
       clientCapabilities.roots?.listChanged == true;
 
   @override
-  FutureOr<ServerCapabilities> initialize(
-    MCPServerInitialization initialization,
-  ) {
+  FutureOr<void> initialize(MCPServerInitialization initialization) {
     initialized.then((_) async {
       if (!supportsRoots) {
         log(