+++ pkgs/dart_mcp/test/api/prompts_test.dart
@@ -81,9 +81,7 @@ final class TestMCPServerWithPrompts extends TestMCPServer with PromptsSupport {
   TestMCPServerWithPrompts(super.channel);
 
   @override
-  FutureOr<ServerCapabilities> initialize(
-    MCPServerInitialization initialization,
-  ) {
+  FutureOr<void> initialize(MCPServerInitialization initialization) {
     addPrompt(greeting, _greetingPrompt);
     return super.initialize(initialization);
   }